Unify download.py and downloadSoftware.py.

+++ scripts/download.py
@@ -1,79 +1,11 @@
+# Download classes from the HITO SPARQL endpoint
+# Not using the CSV2RDF tables because some products are not there but included in the base hito.ttl file.
+# Those need to be removed from hito.ttl afterwards.
+
 import requests
 import csv
 import os
-
-outputBase = "attribute"
-os.makedirs(outputBase,0o777,True)
-
-standard = {
-    "query": '''SELECT REPLACE(STR(?uri),"http://hitontology.eu/ontology/","") as ?suffix
-        STR(SAMPLE(?label)) AS ?label
-        STR(SAMPLE(?comment)) AS ?comment
-        GROUP_CONCAT(?source;separator="|") AS ?sources
- {
-  ?uri a hito:Interoperability;
-          rdfs:label ?label.
-
-  OPTIONAL {?uri <http://purl.org/dc/terms/source> ?source.}
-  OPTIONAL {?uri rdfs:comment ?z.}
-}''',
-    "endpoint": "https://hitontology.eu/sparql",
-    "table": "InteroperabilityStandard",
-    "fields": "(suffix, label, comment, sourceuris)"
-}
-
-language = {
-    "query": '''SELECT REPLACE(STR(?uri),"http://dbpedia.org/resource/","") as ?suffix STR(SAMPLE(?label)) AS ?label
-{
- ?uri a dbo:Language;
-      rdfs:label ?label;
-      dbo:iso6391Code [].
- FILTER(LANGMATCHES(LANG(?label),"en")||LANGMATCHES(LANG(?label),""))
-}''',
-    "endpoint": "https://dbpedia.org/sparql",
-    "table": "Language",
-    "fields": "(suffix, label)"
-}
-
-#  SWO is uploaded to the HITO endpoint, they are (transitive) subclasses, not instances
-license = {
-    "query": '''PREFIX swo: <http://www.ebi.ac.uk/swo/>
-SELECT REPLACE(STR(?uri),"http://www.ebi.ac.uk/swo/license/","") as ?suffix STR(SAMPLE(?label)) AS ?label
-#FROM <http://www.ebi.ac.uk/swo/swo.owl/1.7>
-{
- ?uri rdfs:subClassOf+ swo:SWO_0000002;
-      rdfs:label ?label.
- FILTER(LANGMATCHES(LANG(?label),"en")||LANGMATCHES(LANG(?label),""))
-}''',
-    "endpoint": "https://hitontology.eu/sparql",
-    "table": "License",
-    "fields": "(suffix, label)"
-}
-
-programmingLanguage = {
-    "query": '''SELECT REPLACE(STR(?uri),"http://dbpedia.org/resource/","") as ?suffix STR(SAMPLE(?label)) AS ?label
-{
- ?uri a yago:WikicatProgrammingLanguages ;
-      rdfs:label ?label.
- FILTER(LANGMATCHES(LANG(?label),"en")||LANGMATCHES(LANG(?label),""))
-}''',
-    "endpoint": "https://dbpedia.org/sparql",
-    "table": "ProgrammingLanguage",
-    "fields": "(suffix, label)"
-}
-
-operatingSystem = {
-    "query": '''SELECT REPLACE(STR(?uri),"http://dbpedia.org/resource/","") as ?suffix STR(SAMPLE(?label)) AS ?label
-{
- ?uri a hito:OperatingSystem ;
-      rdfs:label ?label.
- FILTER(LANGMATCHES(LANG(?label),"en")||LANGMATCHES(LANG(?label),""))
-}''',
-    "endpoint": "https://hitontology.eu/sparql",
-    "table": "OperatingSystem",
-    "fields": "(suffix, label)"
-}
-
+from classes import classes
 
 def valueMap(value,isArray):
     if(isArray):
@@ -83,25 +15,34 @@ def valueMap(value,isArray):
         return 'NULL'
     return "E'"+value.replace("'","\\'")+"'" # escape single quotes, add quotes for SQL
 
-def insert(values):
+def insert(values,arrayfields):
     mapped = []
     for i in range(len(values)):
-        mapped.append(valueMap(values[i],i==3)) # there can be more than one source, it is at position 3 counting from 0 
+        mapped.append(valueMap(values[i],i in arrayfields))
     s = ",".join(mapped) 
     return "("+s+")"
 
-classes = [standard,language,license,programmingLanguage,operatingSystem]
-
 for clazz in classes:
+    print("Downloading "+clazz["table"])
     filename=clazz['table']+".sql"
-    output=open(outputBase+"/"+filename, "w")
-    output.write("DELETE FROM "+clazz['table']+";\n")
-    output.write("INSERT INTO "+clazz['table']+clazz['fields']+" VALUES"+'\n')
     parameters = {"query": clazz["query"], "format": "text/tab-separated-values"}
     resp = requests.get(clazz["endpoint"],params=parameters)
+    if(resp.status_code!=200):
+        print("Error with SPARQL query :\n"+resp.text)
+        continue
     readCSV = csv.reader(resp.text.splitlines(), delimiter='\t')
     next(readCSV, None) # skip CSV header
-    content = ",\n".join(map(lambda line: insert(line), readCSV))
-    output.write(content)
-    output.write(";")
-    output.close()
+    content = ",\n".join(map(lambda line: insert(line,clazz["arrayfields"]), readCSV))
+    if(content == ""):
+        print(f"""No entries found for {clazz["table"]}""") #:\n{clazz["query"]}""")
+    else:
+        outputBase = clazz["folder"]
+        os.makedirs(outputBase,0o777,True)
+        output=open(outputBase+"/"+filename, "w")
+        output.write(f"\echo Importing {clazz['table']} from {clazz['endpoint']} \n")
+        output.write("DELETE FROM "+clazz['table']+";\n")
+        output.write("INSERT INTO "+clazz['table']+clazz['fields']+" VALUES"+'\n')
+        output.write(content)
+        output.write("ON CONFLICT DO NOTHING") # skip duplicates instead of cancelling, only for testing
+        output.write(";")
+        output.close()
